+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.9.0 - May 16st, 2018
+
+* Add the Page operator: filters on n tuples according to an ordering
+  and a given page size and page index. No optimization implemented yet.
+
 # 0.8.0 - May 16st, 2018
 
 * Add the Group operator: groups some attributes of the operand as a